metro-ops/requirements.txt
requests
python-dotenv
pandas
supabase
